fix(fs-indexer): resolve Linux compilation issues and rustc ICE
- Work around rustc 1.94/1.96 ICE in mir_borrowck by isolating
  inotify.watches().add() into a separate add_single_watch() helper
- Fix jwalk FileType API: use dir_entry.file_type().is_symlink()
  directly (Linux returns FileType, not Option<FileType>)
- Fix unnecessary i64 cast in detect.rs statfs magic number
- Move chrono::Utc import to test module in enrich.rs (unused in lib)
- Fix clippy manual_flatten in collect_watch_dirs

All 34 unit tests pass on WSL2 Linux, clippy clean on both platforms.

+++ crates/fs-indexer/src/platform/linux/detect.rs
@@ -98,7 +98,7 @@ fn map_fstype(fs_type: &str) -> FilesystemKind {
 pub fn is_pseudo_filesystem(path: &Path) -> bool {
     match nix::sys::statfs::statfs(path) {
         Ok(stat) => {
-            let magic = stat.filesystem_type().0 as i64;
+            let magic = stat.filesystem_type().0;
             matches!(
                 magic,
                 PROC_SUPER_MAGIC
